# api-doc/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "neit Api Documentation",
    "description": "neit Api Documentation",
    "license": {
      "name": "Specify your license",
      "url": "https://yourlicense.com"
    },
    "x-logo": {
      "url": "logo.png",
      "description": "neit Logo",
      "altText": "neit Logo",
      "backgroundColor": "#483D8B"
    },
    "contact": {
      "name": "neit API Support",
      "url": "https://neit.com/info/support",
      "email": "support@neit.com"
    },
    "termsOfService": "https://neit.com/info/terms_and_conditions"
  },
  "servers": [
    {
      "url": "http://localhost:8060/api",
      "description": "Api base url"
    }
  ],
  "paths": {
...
    "/reclutamiento": {
      "get": {
        "summary": "List Reclutamiento",
        "description": "Reclutamiento",
        "operationId": "ListReclutamiento",
        
        "tags": [
          "Reclutamiento"
        ],
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "description": "Filter table records by search value",
            "required": false,
            "type": "string",
            "schema": {
              "$ref": "#/components/schemas/SearchQuery"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of records to select. Default 20",
            "required": false,
            "type": "string",
            "schema": {
              "$ref": "#/components/schemas/LimitQuery"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Current page record",
            "required": false,
            "type": "string",
            "schema": {
              "$ref": "#/components/schemas/PageNumberQuery"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Request Completed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListReclutamiento"
                }
              }
            }
          },
          "default": {
            "description": "Unable to complete request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/reclutamiento/view/{id}": {
      "get": {
        "summary": "View Reclutamiento",
        "description": "Reclutamiento Details",
        "operationId": "ViewReclutamiento",
        
        "tags": [
          "Reclutamiento"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Select record by ID",
            "required": true,
            "type": "string",
            "schema": {
              "$ref": "#/components/schemas/RecordID"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Request Completed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ViewReclutamiento"
                }
              }
            }
          },
          "default": {
            "description": "Unable to complete request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/reclutamiento/add": {
      "post": {
        "summary": "Add Reclutamiento",
        "description": "Add New Reclutamiento",
        "operationId": "AddReclutamiento",
        
        "tags": [
          "Reclutamiento"
        ],
        
        "requestBody": {
          "description": "Add Reclutamiento",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddReclutamiento"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Request Completed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ViewReclutamiento"
                }
              }
            }
          },
          "default": {
            "description": "Unable to complete request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/reclutamiento/edit/{id}": {
      "put": {
        "summary": "Edit Reclutamiento",
        "description": "Edit Reclutamiento",
        "operationId": "EditReclutamiento",
        
        "tags": [
          "Reclutamiento"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Select record by ID",
            "required": true,
            "type": "string",
            "schema": {
              "$ref": "#/components/schemas/RecordID"
            }
          }
        ],
        "requestBody": {
          "description": "Edit Reclutamiento",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditReclutamiento"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Request Completed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EditReclutamiento"
                }
              }
            }
          },
          "default": {
            "description": "Unable to complete request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/reclutamiento/delete/{id}": {
      "delete": {
        "summary": "Delete Reclutamiento",
        "description": "Reclutamiento",
        "operationId": "DeleteReclutamiento",
        
        "tags": [
          "Reclutamiento"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Select record by ID",
            "required": true,
            "type": "string",
            "schema": {
              "$ref": "#/components/schemas/RecordID"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Request Completed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "default": {
            "description": "Unable to complete request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ListReclutamiento": {
        "type": "object",
        "properties": {
          "posiciones": {
            "description": "Posiciones",
            "type": "string"
          },
          "status": {
            "description": "Status",
            "type": "string"
          },
          "descripcion": {
            "description": "Descripcion",
            "type": "string"
          },
          "titulo": {
            "description": "Titulo",
            "type": "string"
          },
          "salario": {
            "description": "Salario",
            "type": "string"
          },
          "tipo": {
            "description": "Tipo",
            "type": "string"
          },
          "fecha": {
            "description": "Fecha",
            "type": "string"
          },
          "telefono": {
            "description": "Telefono",
            "type": "string"
          },
          "correo": {
            "description": "Correo",
            "type": "string"
          }
        }
      },
      "ViewReclutamiento": {
        "type": "object",
        "properties": {
          "posiciones": {
            "description": "Posiciones",
            "type": "string"
          },
          "status": {
            "description": "Status",
            "type": "string"
          },
          "descripcion": {
            "description": "Descripcion",
            "type": "string"
          },
          "titulo": {
            "description": "Titulo",
            "type": "string"
          },
          "salario": {
            "description": "Salario",
            "type": "string"
          },
          "tipo": {
            "description": "Tipo",
            "type": "string"
          },
          "fecha": {
            "description": "Fecha",
            "type": "string"
          },
          "telefono": {
            "description": "Telefono",
            "type": "string"
          },
          "correo": {
            "description": "Correo",
            "type": "string"
          }
        }
      },
      "AddReclutamiento": {
        "type": "object",
        "properties": {
          "posiciones": {
            "description": "Posiciones",
            "type": "string"
          },
          "status": {
            "description": "Status",
            "type": "string"
          },
          "descripcion": {
            "description": "Descripcion",
            "type": "string"
          },
          "titulo": {
            "description": "Titulo",
            "type": "string"
          },
          "salario": {
            "description": "Salario",
            "type": "string"
          },
          "tipo": {
            "description": "Tipo",
            "type": "string"
          },
          "fecha": {
            "description": "Fecha",
            "type": "string"
          },
          "telefono": {
            "description": "Telefono",
            "type": "string"
          },
          "correo": {
            "description": "Correo",
            "type": "string"
          }
        }
      },
      "EditReclutamiento": {
        "type": "object",
        "properties": {
          "posiciones": {
            "description": "Posiciones",
            "type": "string"
          },
          "status": {
            "description": "Status",
            "type": "string"
          },
          "descripcion": {
            "description": "Descripcion",
            "type": "string"
          },
          "titulo": {
            "description": "Titulo",
            "type": "string"
          },
          "salario": {
            "description": "Salario",
            "type": "string"
          },
          "tipo": {
            "description": "Tipo",
            "type": "string"
          },
          "fecha": {
            "description": "Fecha",
            "type": "string"
          },
          "telefono": {
            "description": "Telefono",
            "type": "string"
          },
          "correo": {
            "description": "Correo",
            "type": "string"
          }
        }
      },
...
      "SearchQuery": {
        "type": "string",
        "properties": {
          "search": {
            "description": "Search value",
            "type": "string"
          }
        }
      },
...
      "LimitQuery": {
        "type": "string",
        "properties": {
          "limit": {
            "description": "Maximum number of records to select",
            "type": "integer"
          }
        }
      },
      "PageNumberQuery": {
        "type": "string",
        "properties": {
          "pagenumber": {
            "description": "Current page number",
            "type": "integer"
          }
        }
      },
      "RecordID": {
        "type": "string",
        "properties": {
          "id": {
            "description": "Record ID",
            "type": "string"
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "code": {
            "description": "Http Status Code",
            "type": "integer"
          },
          "message": {
            "description": "Unable to complete request",
            "type": "string"
          }
        }
      }
    }
  }
}